Grab 'talk' events before placing the call.
'talkdetected' is fired whenever the gateway receives audio from
the Flash peer. 'notalkdetected' is fired if no sound activity has
been detected during a period of time that can be configured on
the gateway (15 seconds as of now).

Very useful if we want to make sure that the microphone is properly
set up from the gateway side.