Added new option to create ConferenceRooms.
ConferenceRooms can start in text + audio mode or in text only. In the
latter case, we don't call getUserMedia. If the ConferenceRoom object
was created in text mode only, we need a new function to start the video
in audio mode, probably similar to the funciton we currently have to
join the room in video mode.